requirements.txt
apache-airflow==2.10.2
apache-airflow-providers-http==4.10.0
requests==2.32.3
